+++ Tests/Runtime/Scripts/EnumUtilityTests.cs
@@ -13,9 +13,12 @@ internal class EnumUtilityTests {
     public void ConvertEnumToValueList() {
         int numElements = m_enumValues.Count;
         NUnit.Framework.Assert.Greater(numElements,0);
-        for (int i = 0; i < numElements; ++i) {
-            Assert.AreEqual(m_enumValues[i], (TestEnum)(i));
-        }
+        
+        Assert.AreEqual(m_enumValues[0], (TestEnum) INSPECTOR_KEY_0);
+        Assert.AreEqual(m_enumValues[1], (TestEnum) INSPECTOR_KEY_1);
+        Assert.AreEqual(m_enumValues[2], (TestEnum) INSPECTOR_KEY_2);
+        Assert.AreEqual(m_enumValues[3], (TestEnum) INSPECTOR_KEY_3);
+        Assert.AreEqual(m_enumValues[4], (TestEnum) INSPECTOR_KEY_4);
     }
 
     [Test]
@@ -31,22 +34,40 @@ internal class EnumUtilityTests {
         Assert.AreEqual(names[4], inspectorNames[4]);
     }
     
+    [Test]
+    public void ConvertEnumToInspectorDictionary() {
+        Dictionary<int, string> inspectorNameDictionary = EnumUtility.ToInspectorNameDictionary(typeof(TestEnum));
+        string[]                names                   = Enum.GetNames(typeof(TestEnum));
+        Assert.AreEqual(names.Length, inspectorNameDictionary.Count);
+                    
+        Assert.AreEqual(INSPECTOR_NAME_0, inspectorNameDictionary[INSPECTOR_KEY_0]);
+        Assert.AreEqual(INSPECTOR_NAME_1, inspectorNameDictionary[INSPECTOR_KEY_1]);
+        Assert.AreEqual(INSPECTOR_NAME_2, inspectorNameDictionary[INSPECTOR_KEY_2]);
+        Assert.AreEqual(INSPECTOR_NAME_3, inspectorNameDictionary[INSPECTOR_KEY_3]);
+        Assert.AreEqual(names[4], inspectorNameDictionary[INSPECTOR_KEY_4]);
+    }
 
-//----------------------------------------------------------------------------------------------------------------------
+    
+//--------------------------------------------------------------------------------------------------------------------------------------------------------------
     
     enum TestEnum {
-        [InspectorName(INSPECTOR_NAME_0)] FOO = 0,
-        [InspectorName(INSPECTOR_NAME_1)] BAR,
-        [InspectorName(INSPECTOR_NAME_2)] FOO_BAR,
-        [InspectorName(INSPECTOR_NAME_3)] GOO,
-        HOGE,        
+        [InspectorName(INSPECTOR_NAME_0)] Foo     = INSPECTOR_KEY_0,
+        [InspectorName(INSPECTOR_NAME_1)] Bar     = INSPECTOR_KEY_1,
+        [InspectorName(INSPECTOR_NAME_2)] Foo_Bar = INSPECTOR_KEY_2,
+        [InspectorName(INSPECTOR_NAME_3)] Goo     = INSPECTOR_KEY_3,
+        Hoge = INSPECTOR_KEY_4,
     };
     
     const string INSPECTOR_NAME_0 ="foo";
     const string INSPECTOR_NAME_1 ="bar";
     const string INSPECTOR_NAME_2 ="Foo_Bar";
     const string INSPECTOR_NAME_3 ="GoO";
-    
+
+    const int INSPECTOR_KEY_0 = 0;
+    const int INSPECTOR_KEY_1 = 10;
+    const int INSPECTOR_KEY_2 = 20;
+    const int INSPECTOR_KEY_3 = 30;
+    const int INSPECTOR_KEY_4 = 40;
     
     private readonly List<TestEnum> m_enumValues = EnumUtility.ToValueList<TestEnum>();